feat: v3.0.0 - Package restructure (Phase 2)
Decompose monolithic scripts/detect.py (2736 lines) into prompt_guard/
package with focused modules:

- models.py: Severity, Action, DetectionResult, SanitizeResult
- patterns.py: 500+ regex pattern definitions (pure data)
- normalizer.py: HOMOGLYPHS dict + normalize()
- decoder.py: decode_all() + detect_base64()
- scanner.py: scan_text_for_patterns()
- engine.py: PromptGuard class (analyze, config, rate_limit, canary)
- output.py: scan_output() + sanitize_output() (enterprise DLP)
- logging_utils.py: log_detection(), log_detection_json(), report_to_hivefence()
- cli.py: main() CLI entry point

scripts/ retained as backward-compat shims with DeprecationWarning.
All 121 tests pass. No behavioral changes.
