+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "prompt-guard"
-version = "2.8.2"
+version = "3.0.0"
 description = "Prompt injection defense for any LLM agent — multi-language, context-aware, severity-scored detection with enterprise DLP."
 readme = "README.md"
 license = {text = "MIT"}
@@ -67,10 +67,10 @@ Changelog = "https://github.com/seojoonkim/prompt-guard/blob/main/CHANGELOG.md"
 
 [project.scripts]
 # CLI entry point: `prompt-guard "message to scan"`
-prompt-guard = "scripts.detect:main"
+prompt-guard = "prompt_guard.cli:main"
 
 [tool.setuptools.packages.find]
-include = ["scripts*", "tests*"]
+include = ["prompt_guard*", "scripts*", "tests*"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
