+++ tests/test_detect.py
@@ -29,7 +29,7 @@ from pathlib import Path
 
 # Add project root to path
 sys.path.insert(0, str(Path(__file__).parent.parent))
-from scripts.detect import (
+from prompt_guard import (
     PromptGuard,
     Severity,
     Action,
